pyproject.toml
[project]
name = "livraria-2023"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.1",
    "djangorestframework-simplejwt>=5.3.1",
    "setuptools>=69.2.0",
    "python-magic>=0.4.27",
    "Pillow>=10.2.0",
    "drf-spectacular>=0.27.1",
    "python-dotenv>=1.0.1",
    "netifaces>=0.11.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "django-cloudinary-storage>=0.3.0",
    "django-filter>=24.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[tool.pdm]
distribution = false

[tool.black]
line-length = 120

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.13.2",
    "black>=24.3.0",
]

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver"
dev_ip = "python manage.py runserver 0.0.0.0:19003"
